requirements.txt
absl-py==0.5.0
alabaster==0.7.12
allennlp==0.8.3
asn1crypto==0.24.0
astor==0.7.1
astroid==2.1.0
atomicwrites==1.3.0
attrs==19.1.0
aws-xray-sdk==0.95
awscli==1.16.135
Babel==2.6.0
backcall==0.1.0
bleach==3.1.0
boto==2.49.0
boto3==1.9.88
botocore==1.12.88
bpemb==0.2.11
bz2file==0.98
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.3.9
conllu==0.11
cryptography==2.6.1
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
decorator==4.3.2
Deprecated==1.2.4
dill==0.2.9
docker==3.7.2
docker-pycreds==0.4.0
docutils==0.14
ecdsa==0.13
editdistance==0.5.3
en-core-web-sm==2.0.0
entrypoints==0.3
flair==0.4.1
flaky==3.5.3
Flask==1.0.2
Flask-Cors==3.0.7
ftfy==5.5.1
future==0.17.1
gast==0.2.0
gensim==3.4.0
gevent==1.4.0
greenlet==0.4.15
grpcio==1.15.0
h5py==2.9.0
hyperopt==0.1.1
idna==2.8
imagesize==1.1.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
jmespath==0.9.3
jsondiff==1.1.1
jsonpickle==1.1
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
mock==2.0.0
more-itertools==7.0.0
moto==1.3.7
mpld3==0.3
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.2
neuralcoref==3.1
nltk==3.3
notebook==5.7.4
numpy==1.16.2
numpydoc==0.8.0
olefile==0.46
overrides==1.9
packaging==19.0
pandas==0.24.1
pandocfilters==1.4.2
parsimonious==0.8.1
parso==0.3.2
patsy==0.5.1
pbr==5.1.3
pickleshare==0.7.5
Pillow==5.4.1
plac==0.9.6
plotly==3.6.1
pluggy==0.9.0
preshed==2.0.1
prometheus-client==0.5.0
prompt-toolkit==2.0.8
protobuf==3.6.1
py==1.8.0
pyaml==18.11.0
pyasn1==0.4.5
pycparser==2.19
pycryptodome==3.8.0
Pygments==2.3.1
pylint==2.2.2
pymongo==3.7.2
pyparsing==2.3.1
pypiwin32==223
pyreadline==2.1
pytest==4.4.0
python-dateutil==2.7.5
python-igraph==0.7.1.post6
python-jose==2.0.2
pytorch-pretrained-bert==0.6.1
pytz==2018.9
pywin32==224
pywinpty==0.5.5
PyYAML==3.13
pyzmq==17.1.2
regex==2018.1.10
requests==2.21.0
responses==0.10.6
retrying==1.3.3
rope==0.11.0
rsa==3.4.2
s3transfer==0.1.13
scikit-learn==0.20.2
scipy==1.2.0
seaborn==0.9.0
segtok==1.5.7
Send2Trash==1.5.0
sentencepiece==0.1.81
six==1.11.0
sklearn==0.0
smart-open==1.8.0
snowballstemmer==1.2.1
spacy==2.0.18
spacy-wordnet==0.0.4
Sphinx==2.0.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.1
sqlitedict==1.6.0
sqlparse==0.3.0
statsmodels==0.9.0
syntok==1.1.0
tensorboard==1.10.0
tensorboardX==1.6
tensorflow==1.10.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
thinc==6.12.1
toolz==0.9.0
torch==1.0.0
torchvision==0.2.1
tornado==5.1.1
tqdm==4.26.0
traitlets==4.3.2
typed-ast==1.1.0
typing==3.6.4
ujson==1.35
Unidecode==1.0.23
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.14.1
wincertstore==0.2
word2number==1.1
wrapt==1.10.11
xmltodict==0.12.0
